src/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
soupsieve==2.5
urllib3==2.0.5
wanakana-python==1.2.2
